+++ src/main/java/net/minecraft/client/resources/sounds/MinecartSoundInstance.java
@@ -24,9 +24,9 @@ public class MinecartSoundInstance extends AbstractTickableSoundInstance {
         this.looping = true;
         this.delay = 0;
         this.volume = 0.0F;
-        this.x = (float)minecart.getX();
-        this.y = (float)minecart.getY();
-        this.z = (float)minecart.getZ();
+        this.x = minecart.getX();
+        this.y = minecart.getY();
+        this.z = minecart.getZ();
     }
 
     @Override
@@ -44,9 +44,9 @@ public class MinecartSoundInstance extends AbstractTickableSoundInstance {
         if (this.minecart.isRemoved()) {
             this.stop();
         } else {
-            this.x = (float)this.minecart.getX();
-            this.y = (float)this.minecart.getY();
-            this.z = (float)this.minecart.getZ();
+            this.x = this.minecart.getX();
+            this.y = this.minecart.getY();
+            this.z = this.minecart.getZ();
             float speed = (float)this.minecart.getDeltaMovement().horizontalDistance();
             boolean offRail = !this.minecart.isOnRails() && this.minecart.getBehavior() instanceof NewMinecartBehavior;
             if (speed >= 0.01F && this.minecart.level().tickRateManager().runsNormally() && !offRail) {
